# apps/mobile/lib/generated/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cal Tracker API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          }
        },
        "required": [
          "email",
          "password",
          "displayName"
        ],
        "additionalProperties": false
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "password"
        ],
        "additionalProperties": false
      },
      "RefreshRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 32
          }
        },
        "required": [
          "refreshToken"
        ],
        "additionalProperties": false
      },
...
      "TokenPair": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "email": {
                "type": "string",
                "format": "email"
              },
              "displayName": {
                "type": "string"
              },
              "trustedModeEnabled": {
                "type": "boolean"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": [
              "id",
              "email",
              "displayName",
              "trustedModeEnabled",
              "createdAt"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresAt",
          "user"
        ],
        "additionalProperties": false
      },
      "SettingsUpdate": {
        "type": "object",
        "properties": {
          "trustedModeEnabled": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "GoalsUpdate": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "calories": {
            "type": "integer",
            "minimum": 800,
            "maximum": 10000
          },
          "hydrationGoalGlasses": {
            "type": "integer",
            "minimum": 1,
            "maximum": 40
          }
        },
        "additionalProperties": false
      },
...
      "ExecuteActionRequest": {
        "type": "object",
        "properties": {
          "input": {
            "default": {}
          },
          "source": {
            "type": "string",
            "enum": [
              "flutter",
              "internal_agent",
              "android_appfunctions",
              "ios_appintents",
              "rest"
            ],
            "default": "rest"
          }
        },
        "additionalProperties": false
      },
      "ExecuteActionResponse": {
        "type": "object",
        "properties": {
          "actionCallId": {
            "type": "string",
            "format": "uuid"
          },
          "confirmationRequired": {
            "type": "boolean"
          },
          "output": {}
        },
        "required": [
          "actionCallId",
          "confirmationRequired"
        ],
        "additionalProperties": false
      },
      "AgentRunRequest": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1
          },
          "source": {
            "type": "string",
            "enum": [
              "flutter",
              "ios_appintents",
              "android_appfunctions"
            ],
            "default": "flutter"
          }
        },
        "required": [
          "text"
        ],
        "additionalProperties": false
      },
...
      "TranscriptionResponse": {
        "type": "object",
        "properties": {
          "transcript": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "required": [
          "transcript",
          "provider",
          "model",
          "traceId"
        ],
        "additionalProperties": false
      },
...
      "propose_meal_log_input": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "text"
        ],
        "additionalProperties": false
      },
...
      "correct_meal_input": {
        "type": "object",
        "properties": {
          "mealId": {
            "type": "string",
            "format": "uuid"
          },
          "proposalId": {
            "type": "string",
            "format": "uuid"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "name": {
                  "type": "string",
                  "minLength": 1
                },
                "quantity": {
                  "type": "number",
                  "exclusiveMinimum": 0
                },
                "unit": {
                  "type": "string",
                  "minLength": 1
                },
                "calories": {
                  "type": "integer",
                  "minimum": 0
                },
                "proteinGrams": {
                  "type": "number",
                  "minimum": 0
                },
                "carbsGrams": {
                  "type": "number",
                  "minimum": 0
                },
                "fatGrams": {
                  "type": "number",
                  "minimum": 0
                },
                "source": {
                  "type": "string",
                  "default": "backend_estimate"
                },
                "originalText": {
                  "type": "string"
                },
                "canonicalName": {
                  "type": "string"
                },
                "externalSource": {
                  "type": "string"
                },
                "externalId": {
                  "type": "string"
                },
                "sourceUrl": {
                  "type": "string",
                  "format": "uri"
                },
                "license": {
                  "type": "string"
                },
                "confidence": {
                  "type": "number",
                  "minimum": 0,
                  "maximum": 1
                },
                "needsReview": {
                  "type": "boolean"
                },
                "resolvedGrams": {
                  "type": "number",
                  "exclusiveMinimum": 0
                },
                "portionDescription": {
                  "type": "string"
                }
              },
              "required": [
                "name",
                "quantity",
                "unit",
                "calories",
                "proteinGrams",
                "carbsGrams",
                "fatGrams"
              ],
              "additionalProperties": false
            },
            "minItems": 1
          }
        },
        "required": [
          "items"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/v1/auth/register": {
      "post": {
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token pair",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token pair",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "refresh",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token pair",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/me": {
      "get": {
        "operationId": "getMe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user"
          }
        }
      }
    },
    "/v1/settings": {
      "put": {
        "operationId": "updateSettings",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated settings"
          }
        }
      }
    },
    "/v1/goals": {
      "put": {
        "operationId": "updateGoals",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated daily goals",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/actions": {
      "get": {
        "operationId": "listActions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Action metadata"
          }
        }
      }
    },
    "/v1/actions/{actionId}/execute": {
      "post": {
        "operationId": "executeAction",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "actionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Action result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteActionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/agent/runs": {
      "post": {
        "operationId": "runAgent",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stt/transcriptions": {
      "post": {
        "operationId": "transcribeAudio",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "audio": {
                    "type": "string",
                    "format": "binary"
                  },
                  "source": {
                    "type": "string"
                  }
                },
                "required": [
                  "audio"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transcript",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/meals/proposals": {
      "post": {
        "operationId": "createMealProposal",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/propose_meal_log_input"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Proposal action result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteActionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/meals/proposals/{proposalId}/commit": {
      "post": {
        "operationId": "commitMealProposal",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "proposalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Commit action result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteActionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/meals/{mealId}/correct": {
      "post": {
        "operationId": "correctMeal",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "mealId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/correct_meal_input"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Correction action result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteActionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/meals/{mealId}": {
      "delete": {
        "operationId": "deleteMeal",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "mealId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "confirmationToken",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete action result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteActionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/summary/daily": {
      "get": {
        "operationId": "getDailySummary",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daily summary action result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteActionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/meals": {
      "get": {
        "operationId": "getMealHistory",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Meal history action result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteActionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/meal-templates": {
      "get": {
        "operationId": "getMealTemplates",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Template action result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteActionResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createMealTemplate",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/create_meal_template_input"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Template action result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteActionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/meal-templates/{templateId}": {
      "put": {
        "operationId": "updateMealTemplate",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/update_meal_template_input"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Template action result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteActionResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteMealTemplate",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Template delete action result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteActionResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}